requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
requests
azure-storage-blob
azure-core
azure-ai-formrecognizer
azure-ai-documentintelligence
openai
markdown
tiktoken
langchain
pyodbc
pandas
pyarrow
typing
